Hello world yolo segmentation for QC
