+++ autotomeqc/config/config_loader.py
@@ -2,6 +2,12 @@
 
 import yaml
 from pathlib import Path
+import os
+
+
+TEST_IMG_DIR = Path(r"C:\Users\hanna.lee\Documents\01_lasso\001_data\Training_data_segmask\Trained\qc3_postpickup_20251027_cropped")
+TEST_OUT_DIR = Path(r"C:\Users\hanna.lee\Documents\01_lasso\001_data\Training_data_segmask\Trained\qc3_postpickup_20251027_cropped_out")
+os.makedirs(TEST_OUT_DIR, exist_ok=True)
 
 def load_app_config(filename: str = 'yolo-config.yaml') -> dict:
     """
